feat: restructure folder structure
