+++ src/codegen/functions/function.py
@@ -1,39 +1,42 @@
 """Structure for storing function metadata from scrapigen."""
-from codegen.properties.adaptor_parameter import AdaptorParameter
-from codegen.properties.function_parameters import FunctionsParameter
+from codegen.functions.adaptor_parameter import AdaptorParameter
+from codegen.functions.parameter import Parameter as FunctionParameter
 from codegen.properties.parameter import Parameter
 
 
-class Functions:
+class Function:
     """Structure for storing function metadata from scrapigen."""
 
     def __init__(self, function_name, function_metadata):
         """Structure for storing function metadata from scrapigen."""
         self._function_name = function_name
-        self._c_function_name = function_metadata["c_function_name"]
-        self._description = function_metadata["'description'"]
+        self._description = function_metadata["description"]
         self._is_python_factory = function_metadata.get("is_python_factory", False)
         self._python_class_name = function_metadata["python_class_name"]
         self._calling_convention = function_metadata["calling_convention"]
         self._return_type = function_metadata["returns"]
         if "handle_parameter" in function_metadata:
-            self._handle_parameters = []
-            for name, parameter_data in function_metadata["handle_parameters"].items():
-                self._handle_parameters.append(Parameter(name, parameter_data))
+            self._handle_parameters = Parameter("handle_parameter", function_metadata["handle_parameter"])
 
         if "parameters" in function_metadata:
             self._parameters = []
             for parameter in function_metadata["parameters"].items():
-                self._parameters.append(FunctionsParameter(parameter))
+                self._parameters.append(FunctionParameter(parameter))
 
         if "adaptor_parameter" in function_metadata:
             self._adaptor_parameter = AdaptorParameter(function_metadata("adaptor_parameter"))
+
+        if "cname" in function_metadata:
+            assert function_metadata["cname"].startswith("DAQmx")
+            self._c_function_name = function_metadata["cname"][5:]
+        else:
+            self._c_function_name = function_name
 
     @property
     def function_name(self):
         """str: The name of the function."""
         return self._function_name
-
+    
     @property
     def c_function_name(self):
         """str: The name of the c function to be called when using the function."""
@@ -66,7 +69,7 @@ class Functions:
 
     @property
     def handle_parameters(self):
-        """List of handle parameters: The handle parameter for the instance the function."""
+        """Handle parameter: The handle parameter for the instance the function."""
         return self._handle_parameters
 
     @property
